Refactor Node app structure to modular routes and pages
